change config naming
