+++ src/main/resources/application.yaml
@@ -11,7 +11,8 @@ spring:
       enabled: true
 
 app:
-  route-map:
-    atmCount: 1000
-  collector:
-    count: 5
+  generator:
+    route-map:
+      atmCount: 1000
+    collector:
+      count: 5
